# src/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Swagger Library Manager - OpenAPI 3.0",
        "description": "Esta é a documentação da API para gerenciamento de uma biblioteca",
        "contact": {
            "email": "gabriel.51905@aluno.pr.senac.br"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3333"
        }
    ],
...
    "paths": {
...
        "/books":{
            "post": {
                "tags": ["books"],
                "summary": "Cria um novo Livro",
                "description": "Cria um novo Livro passando suas informações básicas",
                "requestBody": {
                    "content": {
                        "apllication/json": {
                            "schema": {
                                "items":{
                                    "type": "#components/schemas/Book"
                                }
                            },
                            "examples": {
                                "book": {
                                    "value": {
                                        "author": "Machado de Assis",
                                        "title": "Dom Casmurro",
                                        "category": "Romance",
                                        "pages": "346"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Livro Existente" 
                    },
                    "200": {
                        "description": "Livro criado com sucesso!",
                        "content": {
                            "apllication/json": {
                                "schema": {
                                    "type": "object",
                                    "items": {
                                        "type": "#components/schemas/Book"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "get":{
                "tags": ["books"],
                "summary": "Listagem de Livros",
                "description": "Fazer uma lista com todos os Livros cadastrados",
                "requestBody":{
                    "content":{
                        "apllication/json":{
                            "schema":{
                                "items":{
                                    "type":"string"
                                }
                            },
                            "examples":{
                                "book":["books"]
                                }
                            },
                            "responses":{
                                "400":{
                                    "description": "Lista de Livro está vazia"
                                },
                                "200":{
                                    "description": "Lista encontrada!",
                                    "content": {
                                        "apllication/json": {
                                            "schema": {
                                                "type": "object",
                                                "items": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }

            },
            "delete":{
                "tags": ["books"],
                  "summary": "Deletar Livro",
                  "description": "Deleta todas as informações do Livro",
                  "operationId": "deleteBook",
                  "parameters": [{
                      "name": "author, title, category, pages",
                      "in": "path",
                      "description": "Deleta informações sobre o Livro",
                      "required": true,
                      "schema": {
                        "type": "string"
                      }
                    }],
                  "responses": {
                    "400": {
                      "description": "Livro invalido"
                    },
                    "404": {
                      "description": "Livro não encontrado no registro!"
                    }
                  }
                }
        },
...
}
}